runtime maintenance
libcob:
* cconv.c (cob_field_to_string): moved from common.c
* cconv.c (cob_init_cconv), coblocal.h, common.c: added
* cconv.c (cob_toupper, cob_tolower, cob_init_cconv), coblocal.h: included (copy from cobc/reserved.c)
* coblocal.h (enum cob_case_modifier): new enum to select case-adjustment
* cconv.c (cob_field_to_string), coblocal.h: adjusted function signature, allowing for optional change of case during copy and returning the size to process or negative error values in case of "no data to process"
* cconv.c (cob_field_to_string): return error if truncating would be necessary
* reportio.c, intrinsic.c, fileio.c, common.c, call.c: adjusted for new signature of cob_field_to_string
* intrinsic.c (locale_time, cob_intr_locale_date, cob_intr_locale_compare): always use full locale string passed, error handling improved
* common.c (cob_set_locale, cob_get_environment, cob_display_env_value): use local buffer, error handling improved

+++ libcob/fileio.c
@@ -6157,7 +6157,7 @@ cob_pre_open (cob_file *f)
 	} else
 	if (f->assign != NULL
 	 && f->assign->data != NULL) {
-		cob_field_to_string (f->assign, file_open_name, (size_t)COB_FILE_MAX, CCM_NONE);
+		cob_field_to_string (f->assign, file_open_name, COB_FILE_MAX, CCM_NONE);
 	}
 }
 
@@ -6939,7 +6939,7 @@ cob_delete_file (cob_file *f, cob_field *fnstatus)
 	}
 
 	/* Obtain the file name */
-	cob_field_to_string (f->assign, file_open_name, (size_t)COB_FILE_MAX, CCM_NONE);
+	cob_field_to_string (f->assign, file_open_name, COB_FILE_MAX, CCM_NONE);
 	cob_chk_file_mapping ();
 
 	if (f->organization != COB_ORG_INDEXED) {
@@ -8512,9 +8512,9 @@ cob_get_filename_print (cob_file* file, const int show_resolved_name)
 {
 	size_t offset = 0, len;
 	/* Obtain the file name */
-	cob_field_to_string (file->assign, file_open_env, (size_t)COB_FILE_MAX, CCM_NONE);
+	cob_field_to_string (file->assign, file_open_env, COB_FILE_MAX, CCM_NONE);
 	if (show_resolved_name) {
-		strncpy (file_open_name, file_open_env, (size_t)COB_FILE_MAX);
+		strncpy (file_open_name, file_open_env, COB_FILE_MAX);
 		file_open_name[COB_FILE_MAX] = 0;
 		cob_chk_file_mapping ();
 	}
